third_party/WebKit/LayoutTests/fast/scroll-behavior/smooth-scroll/ongoing-smooth-scroll-vertical-rl-anchors.html

<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<style>
  html {
    -webkit-writing-mode: vertical-rl;
  }
  body {
    margin: 0px;
    height: 2000px;
    width: 2000px;
  }
  #changer {
    background-color: #FF7F7F;
    width: 10px;
  }
  #anchor {
    width: 1500px;
    background-color: #84BE6A;
  }
</style>

<div id="changer"></div>
<div id="anchor"></div>

<script>
  var asyncTest = async_test("Verify smooth scroll interaction with scroll anchoring on page with vertical-lr writing mode");

  // The element that will change in height.
  var ch;
  // Initital scroll position.
  var initialX = -10;
  var initialY = 0;
  // Amount to smooth scroll by.
  var userScrollX = -205;
  var userScrollY = 51;
  // Amount to change the height of the element above the viewport.
  var changerX = 100;
  // End position: height of ch + userScroll.
  var endX = -305;
  var endY = 51;

  function scrollListener() {
    if (window.scrollX == endX && window.scrollY == endY) {
      asyncTest.done();
      return;
    }

    if (ch.style.width != "100")
      ch.style.width = changerX;
  }

  window.onload = function() {
    assert_false(!window.internals, 'This test requires internals');
    internals.runtimeFlags.scrollAnchoringEnabled = true;
    // Turn on smooth scrolling.
    internals.settings.setScrollAnimatorEnabled(true);

    ch = document.getElementById("changer");
    window.scrollTo(initialX, initialY);

    asyncTest.step(function() {
      assert_equals(window.scrollX, initialX);
      assert_equals(window.scrollY, initialY);

      // Smooth scroll.
      eventSender.mouseMoveTo(100, 100);
      eventSender.continuousMouseScrollBy(-userScrollX, -userScrollY);

      document.addEventListener("scroll", scrollListener);
    });
  }
</script>
